refactor!: move PdbClient to socket.py
BREAKING CHANGE: break import of PdbClient
