--- pdb_attach/client.pyi
@@ -1,12 +1,0 @@
-from typing import Callable, Tuple, Union
-
-class PdbClient:
-    server_pid: int = ...
-    port: Union[int, str] = ...
-    def __init__(self, pid: int, port: Union[int, str]) -> None: ...
-    def connect(self) -> None: ...
-    def send_cmd(self, cmd: str) -> None: ...
-    send: Callable[[str], None] = ...
-    def recv(self) -> Tuple[str, bool]: ...
-    def send_and_recv(self, cmd: str) -> Tuple[str, bool]: ...
-    def interactive_loop(self) -> None: ...
